requirements.txt
torch>=1.13.0
transformers>=4.30.0
numpy
pandas
matplotlib
tqdm
pyyaml
scikit-learn
accelerate
einops
# Use faiss-gpu
sentence-transformers
 if running on GPU, otherwise faiss-cpu
faiss-gpu
